requirements.txt
streamlit
pandas
plotly
openpyxl
google-generativeai
python-dotenv
reportlab